+++ e2e/pages/sg-list.po.ts
@@ -1,0 +1,28 @@
+import { browser, by, element, protractor } from 'protractor';
+import { CloudstackUiPage } from './app.po';
+
+export class SGList extends CloudstackUiPage {
+  clickSharedTab() {
+    element
+      .all(by.name('viewMode'))
+      .get(2)
+      .click();
+    const EC = protractor.ExpectedConditions;
+    browser.wait(EC.visibilityOf(element(by.css('.entity-card.mat-card'))), 5000);
+  }
+
+  clickPrivateTab() {
+    element
+      .all(by.name('viewMode'))
+      .get(3)
+      .click();
+    const EC = protractor.ExpectedConditions;
+    browser.wait(EC.visibilityOf(element(by.css('.entity-card.mat-card'))), 5000);
+  }
+
+  clickOpenSidebar() {
+    element(by.css('.entity-card.mat-card')).click();
+    const EC = protractor.ExpectedConditions;
+    browser.wait(EC.visibilityOf(element(by.tagName('h4'))), 5000);
+  }
+}