+++ src/app/reducers/vm/redux/vm.reducers.ts
@@ -16,9 +16,9 @@ import { VmCreationState } from '../../../vm/vm-creation/data/vm-creation-state'
 import { VmCreationSecurityGroupData } from '../../../vm/vm-creation/security-group/vm-creation-security-group-data';
 import { notSelectedSshKey } from '../../../vm/vm-creation/ssh-key-selector/ssh-key-selector.component';
 import { noGroup } from '../../../vm/vm-filter/vm-filter.component';
+
 import * as fromAccounts from '../../accounts/redux/accounts.reducers';
 import * as affinityGroupActions from '../../affinity-groups/redux/affinity-groups.actions';
-import * as fromSGroup from '../../security-groups/redux/sg.reducers';
 import * as fromZones from '../../zones/redux/zones.reducers';
 import * as vmActions from './vm.actions';
 
@@ -45,6 +45,7 @@ export interface FormState {
   loaded: boolean;
   showOverlay: boolean;
   deploymentInProgress: boolean;
+  isError: boolean;
   enoughResources: boolean;
   insufficientResources: string[];
   loggerStageList: ProgressLoggerMessage[];
@@ -192,10 +193,7 @@ export function listReducer(state = initialListState, action: vmActions.Actions)
 
 export const getVMsState = createFeatureSelector<VirtualMachineState>('virtualMachines');
 
-export const getVMsEntitiesState = createSelector(
-  getVMsState,
-  state => state.list,
-);
+export const getVMsEntitiesState = createSelector(getVMsState, state => state.list);
 
 export const {
   selectIds,
@@ -204,20 +202,11 @@ export const {
   selectTotal: getVMCount,
 } = adapter.getSelectors(getVMsEntitiesState);
 
-export const isLoading = createSelector(
-  getVMsEntitiesState,
-  state => state.loading,
-);
+export const isLoading = createSelector(getVMsEntitiesState, state => state.loading);
 
-export const isLoaded = createSelector(
-  getVMsEntitiesState,
-  state => state.loaded,
-);
+export const isLoaded = createSelector(getVMsEntitiesState, state => state.loaded);
 
-export const getSelectedId = createSelector(
-  getVMsEntitiesState,
-  state => state.selectedVMId,
-);
+export const getSelectedId = createSelector(getVMsEntitiesState, state => state.selectedVMId);
 
 export const getSelectedVM = createSelector(
   getVMsState,
@@ -230,73 +219,40 @@ export const getSelectedVmAffinityGroups = createSelector(
   vm => vm && vm.affinitygroup,
 );
 
-export const filters = createSelector(
-  getVMsEntitiesState,
-  state => state.filters,
+export const getSelectedVmSecurityGroup = createSelector(
+  getSelectedVM,
+  vm => vm && vm.securitygroup,
 );
+
+export const filters = createSelector(getVMsEntitiesState, state => state.filters);
 
 export const attachmentFilters = createSelector(
   getVMsEntitiesState,
   state => state.attachmentFilters,
 );
 
-export const filterQuery = createSelector(
-  filters,
-  state => state.query,
-);
+export const filterQuery = createSelector(filters, state => state.query);
 
-export const filterSelectedZoneIds = createSelector(
-  filters,
-  state => state.selectedZoneIds,
-);
+export const filterSelectedZoneIds = createSelector(filters, state => state.selectedZoneIds);
 
-export const filterSelectedStates = createSelector(
-  filters,
-  state => state.selectedStates,
-);
+export const filterSelectedStates = createSelector(filters, state => state.selectedStates);
 
-export const filterSelectedGroupNames = createSelector(
-  filters,
-  state => state.selectedGroupNames,
-);
+export const filterSelectedGroupNames = createSelector(filters, state => state.selectedGroupNames);
 
-export const filterSelectedAccountIds = createSelector(
-  filters,
-  state => state.selectedAccountIds,
-);
+export const filterSelectedAccountIds = createSelector(filters, state => state.selectedAccountIds);
 
-export const filterSelectedGroupings = createSelector(
-  filters,
-  state => state.selectedGroupings,
-);
+export const filterSelectedGroupings = createSelector(filters, state => state.selectedGroupings);
 
-export const selectVmGroups = createSelector(
-  selectAll,
-  vms => {
-    const groups = vms.map(vm => vm.group).filter(Boolean);
-    return uniq(groups);
-  },
-);
+export const selectVmGroups = createSelector(selectAll, vms => {
+  const groups = vms.map(vm => vm.group).filter(Boolean);
+  return uniq(groups);
+});
 
-export const getUsingSGVMs = createSelector(
-  selectAll,
-  fromSGroup.getSelectedId,
-  (vms: VirtualMachine[], sGroupId: string) => {
-    const sGroupFilter = (vm: VirtualMachine) =>
-      vm.securitygroup.find(group => group.id === sGroupId);
-    return vms.filter(sGroupFilter);
-  },
-);
+export const getAttachmentVMs = createSelector(selectAll, attachmentFilters, (vms, filter) => {
+  const accountFilter = vm => vm.account === filter.account && vm.domainid === filter.domainId;
 
-export const getAttachmentVMs = createSelector(
-  selectAll,
-  attachmentFilters,
-  (vms, filter) => {
-    const accountFilter = vm => vm.account === filter.account && vm.domainid === filter.domainId;
-
-    return vms.filter(accountFilter);
-  },
-);
+  return vms.filter(accountFilter);
+});
 
 export const selectFilteredVMs = createSelector(
   selectAll,
@@ -366,6 +322,7 @@ export const initialFormState: FormState = {
   loaded: false,
   showOverlay: false,
   deploymentInProgress: false,
+  isError: false,
   enoughResources: false,
   insufficientResources: [],
   loggerStageList: [],
@@ -419,6 +376,7 @@ export function formReducer(
         ...state,
         showOverlay: true,
         deploymentInProgress: true,
+        isError: false,
       };
     }
     case vmActions.VM_DEPLOYMENT_INIT_ACTION_LIST: {
@@ -452,7 +410,12 @@ export function formReducer(
         return message;
       });
 
-      return { ...state, loggerStageList: [...messages], deploymentInProgress: false };
+      return {
+        ...state,
+        loggerStageList: [...messages],
+        deploymentInProgress: false,
+        isError: true,
+      };
     }
     case vmActions.VM_DEPLOYMENT_COPY_TAGS: {
       return {
@@ -470,60 +433,35 @@ export function formReducer(
   }
 }
 
-export const getVmForm = createSelector(
-  getVMsState,
-  state => state.form,
-);
+export const getVmForm = createSelector(getVMsState, state => state.form);
 
-export const getVmFormState = createSelector(
-  getVMsState,
-  state => state.form.state,
-);
+export const getVmFormState = createSelector(getVMsState, state => state.form.state);
 
 export const getVmFormStateAffinityGroup = createSelector(
   getVMsState,
   state => state.form.state.affinityGroup,
 );
 
-export const formIsLoading = createSelector(
-  getVmForm,
-  state => state.loading,
-);
+export const formIsLoading = createSelector(getVmForm, state => state.loading);
 
-export const formIsLoaded = createSelector(
-  getVmForm,
-  state => state.loaded,
-);
+export const formIsLoaded = createSelector(getVmForm, state => state.loaded);
 
-export const enoughResources = createSelector(
-  getVmForm,
-  state => state.enoughResources,
-);
+export const enoughResources = createSelector(getVmForm, state => state.enoughResources);
 
 export const insufficientResources = createSelector(
   getVmForm,
   state => state.insufficientResources,
 );
 
-export const deploymentInProgress = createSelector(
-  getVmForm,
-  state => state.deploymentInProgress,
-);
+export const deploymentInProgress = createSelector(getVmForm, state => state.deploymentInProgress);
 
-export const loggerStageList = createSelector(
-  getVmForm,
-  state => state.loggerStageList,
-);
+export const isError = createSelector(getVmForm, state => state.isError);
 
-export const showOverlay = createSelector(
-  getVmForm,
-  state => state.showOverlay,
-);
+export const loggerStageList = createSelector(getVmForm, state => state.loggerStageList);
 
-export const getDeployedVM = createSelector(
-  getVmForm,
-  state => state.deployedVm,
-);
+export const showOverlay = createSelector(getVmForm, state => state.showOverlay);
+
+export const getDeployedVM = createSelector(getVmForm, state => state.deployedVm);
 
 export const getVmCreationZoneId = createSelector(
   getVmFormState,