+++ src/app/root-store/server-data/vm-snapshots/vm-snapshots.effects.ts
@@ -57,6 +57,8 @@ import {
   VmSnapshotActionTypes,
 } from './vm-snapshots.actions';
 import * as vmSnapshotSelectors from './vm-snapshots.selectors';
+import { VmState } from '../../../vm';
+import { configSelectors } from '../../config';
 
 @Injectable()
 export class VmSnapshotsEffects {
@@ -74,8 +76,33 @@ export class VmSnapshotsEffects {
   @Effect()
   createVmSnapshotDialog$: Observable<Action> = this.actions$.pipe(
     ofType<Create>(VmSnapshotActionTypes.Create),
-    exhaustMap(action =>
-      this.matDialog
+    withLatestFrom(
+      this.store.pipe(select(fromVM.getSelectedVM)),
+      this.store.pipe(select(configSelectors.get('vmSnapLimit'))),
+      this.store.pipe(select(vmSnapshotSelectors.getVmSnapshotsNumberForSelectedVm)),
+    ),
+    exhaustMap(([action, selectedVM, vmSnapLimit, snapshotsNumber]) => {
+      if (vmSnapLimit.enable) {
+        if (vmSnapLimit.snapshotsLimit <= snapshotsNumber) {
+          const message = 'ERRORS.VM_SNAPSHOT.LIMIT_EXCEEDED';
+          this.dialogService.showNotificationsOnFail({ message }, message);
+          return of(new CreateCanceled());
+        }
+      }
+
+      if (selectedVM.state === VmState.Stopped) {
+        const message = 'ERRORS.SNAPSHOT.CREATION_UNAVAILABLE_FOR_STOPPED';
+        this.dialogService.showNotificationsOnFail({ message }, message);
+        return of(new CreateCanceled());
+      }
+
+      if (selectedVM.state === VmState.Destroyed) {
+        const message = 'ERRORS.SNAPSHOT.CREATION_UNAVAILABLE_FOR_DELETED';
+        this.dialogService.showNotificationsOnFail({ message }, message);
+        return of(new CreateCanceled());
+      }
+
+      return this.matDialog
         .open(VmSnapshotCreationDialogComponent, { width: '400px', disableClose: true })
         .afterClosed()
         .pipe(
@@ -90,8 +117,8 @@ export class VmSnapshotsEffects {
             }
             return new CreateCanceled();
           }),
-        ),
-    ),
+        );
+    }),
   );
 
   @Effect()